modify the new_sized() constructor to take a gsize param instead of guint
2008-02-02  Jonathon Jongsma  <jjongsma@gnome.org>

	* gio/gbufferedoutputstream.c:
	* gio/gbufferedoutputstream.h: modify the new_sized() constructor to take a
	gsize param instead of guint to match the GBufferedInputStream constructor.



svn path=/trunk/; revision=6448

+++ gio/gbufferedoutputstream.h
@@ -74,7 +74,7 @@ struct _GBufferedOutputStreamClass
 GType           g_buffered_output_stream_get_type  (void) G_GNUC_CONST;
 GOutputStream* g_buffered_output_stream_new             (GOutputStream         *base_stream);
 GOutputStream* g_buffered_output_stream_new_sized       (GOutputStream         *base_stream,
-							 guint                  size);
+							 gsize                  size);
 gsize          g_buffered_output_stream_get_buffer_size (GBufferedOutputStream *stream);
 void           g_buffered_output_stream_set_buffer_size (GBufferedOutputStream *stream,
 							 gsize                  size);
